update project style
