+++ src/assets/style/App.module.css
@@ -1,0 +1,3 @@
+.appContainer {
+  padding-bottom: 4rem;
+}
