+++ src/assets/style/Rocket.module.css
@@ -1,7 +1,9 @@
 .rocketItem {
   display: grid;
   grid-template-columns: 20fr 80fr;
-  margin: 0.5rem 1rem 0.5rem 1rem;
+  margin: 1rem 1rem 0.5rem 1rem;
+  padding-left: 1rem;
+  padding-right: 1rem;
 }
 
 .rocketImageContainer img {
@@ -12,5 +14,18 @@
 .rocketBody {
   display: flex;
   flex-direction: column;
-  padding-left: 0.5rem;
+  padding-left: 1.2rem;
+}
+
+.rocketBody h3 {
+  padding-bottom: 0.5rem;
+}
+
+.btnReserve {
+  margin-top: 0.3rem;
+  width: 25%;
+}
+
+.badge {
+  margin-right: 0.3rem;
 }
