+++ src/assets/style/missions.module.css
@@ -1,61 +1,72 @@
-
-.mission {
-  margin-left: 20%;
+/* .mission {
+  margin-left: 10%;
   border-radius: 4px;
   border: 1px solid #e8e8e8;
   padding: 2rem;
   display: flex;
   flex-wrap: wrap;
   justify-content: space-between;
-  width: 90%; /* Adjusted width to fit content */
-}
+  width: 50%;
+  background: red;
+} */
 
-.mission-description {
-  padding: 0.5em 0;
-}
+/* .mission-description {
+  padding-top: 0.5em;
+  padding-left: 0.5em;
+  padding-right: 0.5em;
+  padding-bottom: 0.5em;
+} */
 
-.mission:nth-child(odd),
-.mission-table tr:nth-child(odd) {
+tbody:nth-child(odd),
+tbody tr:nth-child(odd) {
   background-color: #f0f0f0;
 }
 
-.mission:hover,
+.mission-descriptionContainer p {
+  color: red;
+}
+
+/* .mission:hover,
 .mission-table tr:hover {
   background-color: #f3f3f3;
   border: 1px solid #14b1e0;
-}
+} */
 
 /* Button styles for Mission component */
-.join,
+/* .join,
 .leave {
   padding: 0.3em 0.5em;
   border-radius: 10px;
   cursor: pointer;
-  width: 14%;
-  transition: background-color 0.3s ease; /* Added transition for smoother color change */
-}
+  width: 10%;
+  transition: background-color 0.3s ease;
+} */
 
+/* Added transition for smoother color change */
+/* 
 .join {
   border-color: #949495;
   color: #949495;
   background-color: unset;
-}
+} */
 
-.leave {
+/* .leave {
   border-color: #f14d1b;
   color: #f14d1b;
   background-color: unset;
-}
+} */
 
-.join:hover {
-  background-color: #14b1e0; /* Change to blue on hover */
+/* .join:hover {
+  background-color: #14b1e0;
   color: #fff;
-}
+} */
+/* Change to blue on hover */
 
-.leave:hover {
-  background-color: #f33; /* Change to red on hover */
+/* .leave:hover {
+  background-color: #f33;
   color: #fff;
-}
+} */
+/* Change to red on hover */
 
 /* Additional styles for Mission component */
 .action-buttons {
@@ -90,14 +101,10 @@
   padding: 1em 1em;
 }
 
-.mission-table thead {
-  text-align: center;
-}
-
 /* Additional styles for specific elements */
 
 .mission-status {
-  width: 18%;
+  width: 10%;
 }
 
 .mission-button {
@@ -105,6 +112,10 @@
 }
 
 .mission-name {
-  width: 10%;
-  font-size: 40px;
+  width: 8%;
+  font-size: 1rem;
+}
+
+.btnJoinMission {
+  font-size: 10px;
 }
