+++ src/test/Header.test.js
@@ -5,6 +5,7 @@ import Logo from '../components/Logo';
 import Nav from '../components/Nav';
 
 describe('testing the Header, Logo and Nav Components', () => {
+
   test('Logo component Should contain Image', () => {
     render(<Logo />);
     const img = screen.getByRole('img');
